+++ packages/jest-resolve-dependencies/src/__tests__/dependency_resolver.test.ts
@@ -58,6 +58,17 @@ test('resolves dependencies for existing path', () => {
   ]);
 });
 
+test('includes the mocks of dependencies as dependencies', () => {
+  const resolved = dependencyResolver.resolve(
+    path.resolve(__dirname, '__fixtures__/hasMocked/file.test.js'),
+  );
+
+  expect(resolved).toEqual([
+    expect.stringContaining(path.join('hasMocked', 'file.js')),
+    expect.stringContaining(path.join('hasMocked', '__mocks__', 'file.js')),
+  ]);
+});
+
 test('resolves dependencies for scoped packages', () => {
   const resolved = dependencyResolver.resolve(
     path.resolve(__dirname, '__fixtures__', 'scoped.js'),
@@ -85,6 +96,19 @@ test('resolves inverse dependencies for existing path', () => {
   expect(resolved).toEqual([
     expect.stringContaining(
       path.join('__tests__', '__fixtures__', 'file.test.js'),
+    ),
+  ]);
+});
+
+test('resolves inverse dependencies of mock', () => {
+  const paths = new Set([
+    path.resolve(__dirname, '__fixtures__/hasMocked/__mocks__/file.js'),
+  ]);
+  const resolved = dependencyResolver.resolveInverse(paths, filter);
+
+  expect(resolved).toEqual([
+    expect.stringContaining(
+      path.join('__tests__', '__fixtures__', 'hasMocked', 'file.test.js'),
     ),
   ]);
 });